requirements.txt
click==7.1.2
falcon==2.0.0
pycryptodome==3.9.9
urllib3==1.26.2
Werkzeug==1.0.1
